Customize webpack config
Our project requires some modifications to the webpack config. This change
also removes the duplicate webpack config in `startCompile` and `watchCompile`
by creating a new function `getWebpackConfig`. In future it would be good
to work out how to load this webpack configuration dynamically from the
project folder, however this has proven difficult since the compiler itself
is run through webpack, which attempts to resolve all calls to `require`. The
path of the webpack config cannot be known until runtime, when the user
selects a project folder.