Allow user to manually edit the location of new component source
When generating a new component, this change allows the user to
specify the location of the component source file in the second
step of the wizard.

This change also disabled the ability to generate Reflux actions and
store, as this caused issues and is not yet required for our project.