Added quick property editor to the sidebar panels
This change introduces a 'React Props' panel on the side of the
editor above the style editor panels. Currently the panel displays
any properties that are currently set on the selected component and
allows quick editing of text, numeric, and boolean type properties.
Nested object type properties are not yet supported.