Quick Prop Editor: Disable editing of Object and Array type props
Prior to this change, object and array type props were displayed
as a text box with value [object Object]. Editing this value deleted
the object and replaced it with the string '[object Object]'.
This change disables editing of object and array type props from
the quick prop editor until a solution for editing nested objects
and arrays can be built.