First attempt to fix excessive re-rendering of sidebar panels
Prior to this, every time a value was changed in the sidebar
style panels or react props panels, the component was automatically
deselected and then re-selected, causing the entire sidebar DOM to
be emptied and then re-rendered. This caused many usability issues,
including:
- The sidebar scrolled back to the top every time a change is made
- Clicking the increment/decrement buttons too quickly caused the
  value to jump around

This change solves the above issues, however other issues have
surfaced that were not visible due to the forced re-rendering. These
include:
- The 'isDisabled' state is not updated on style editor components
  unless the user manually deselects and reselects the component
- The selection bound box on the desk frame sometimes does not update
  its bounds in response to a component's size changing
- Changing a component's variant with the 'Load Variant' button loads
  the variant correctly, but the React Props panel is not rendered
  with the updated props